Refactor ImplementationProvider do a different package
